requirements.txt
discord
requests
python-decouple
openai